nautilus-trader/references/requirements.txt
# Core
nautilus_trader>=1.222.0,<1.223.0

# Hyperliquid SDK
hyperliquid-python-sdk>=0.10.0
eth-account>=0.11.0

# Environment
python-dotenv>=1.0.0
